+++ app/Modules/SpaceGame/GeneratingAdapter/Commands/BuilderMacroCommand.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Modules\SpaceGame\IoC\Commands\AdapterBuilding;
+namespace App\Modules\SpaceGame\GeneratingAdapter\Commands;
 
 use App\Modules\SpaceGame\IoC\Commands\ICommand;
