+++ app/Modules/SpaceGame/GeneratingAdapter/Commands/ConcreteMethodCommand.php
@@ -2,12 +2,12 @@
 
 declare(strict_types=1);
 
-namespace App\Modules\SpaceGame\IoC\Commands\AdapterBuilding;
+namespace App\Modules\SpaceGame\GeneratingAdapter\Commands;
 
+use App\Modules\SpaceGame\GeneratingAdapter\ValueObjects\AdapterBuilderVO;
+use App\Modules\SpaceGame\IoC\Commands\ICommand;
 use App\Modules\SpaceGame\IoC\Exceptions\DependencyNotFound;
 use App\Modules\SpaceGame\IoC\IoC;
-use App\Modules\SpaceGame\IoC\Commands\ICommand;
-use App\Modules\SpaceGame\IoC\ValueObjects\AdapterBuilderVO;
 
 final readonly class ConcreteMethodCommand implements ICommand
 {
